change max_delay type to float

+++ 0x01-python_async_function/1-concurrent_coroutines.py
@@ -5,7 +5,7 @@ from typing import List
 wait_random = __import__("0-basic_async_syntax").wait_random
 
 
-async def wait_n(n: int, max_delay: float) -> List[float]:
+async def wait_n(n: int, max_delay: int) -> List[float]:
     """Create coroutines and await their completion upon which
     they will be added to the list to be returned"""
     tasks = [wait_random(max_delay) for _ in range(n)]
